requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1.3
django-discover-runner==1.0
django-role-permissions==3.1.1
django-toolbelt==0.0.1
gunicorn==20.0.4
Pillow==7.2.0
pytz==2019.3
reportlab==3.5.49
selenium==3.141.0
six==1.15.0
sqlparse==0.3.1
static3==0.7.0
urllib3==1.26.6
whitenoise==5.0.1
